+++ plugins/destination/s3/client/spec.go
@@ -2,6 +2,7 @@ package client
 
 import (
 	"fmt"
+	"strings"
 )
 
 type FormatType string
@@ -16,9 +17,19 @@ type Spec struct {
 	Path     string     `json:"path,omitempty"`
 	Format   FormatType `json:"format,omitempty"`
 	NoRotate bool       `json:"no_rotate,omitempty"`
+	Athena   bool       `json:"athena,omitempty"`
 }
 
-func (*Spec) SetDefaults() {}
+func (s *Spec) SetDefaults() {
+	if !strings.Contains(s.Path, PathVarTable) {
+		// for backwards-compatibility, default to given path plus /{{TABLE}}.[format].{{UUID}} if
+		// no {{TABLE}} value is found in the path string
+		s.Path += fmt.Sprintf("/%s.%s", PathVarTable, s.Format)
+		if !s.NoRotate {
+			s.Path += "." + PathVarUUID
+		}
+	}
+}
 
 func (s *Spec) Validate() error {
 	if s.Bucket == "" {
@@ -26,6 +37,9 @@ func (s *Spec) Validate() error {
 	}
 	if s.Path == "" {
 		return fmt.Errorf("path is required")
+	}
+	if s.NoRotate && strings.Contains(s.Path, PathVarUUID) {
+		return fmt.Errorf("path should not contain %s when no_rotate = true", PathVarUUID)
 	}
 	if s.Format == "" {
 		return fmt.Errorf("format is required")