+++ plugins/source/azure/resources/services/redis/caches.go
@@ -3,7 +3,7 @@ package redis
 import (
 	"context"
 
-	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/redis/armredis"
+	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/redis/armredis/v2"
 	"github.com/cloudquery/cloudquery/plugins/source/azure/client"
 	"github.com/cloudquery/plugin-sdk/schema"
 	"github.com/cloudquery/plugin-sdk/transformers"
@@ -17,6 +17,11 @@ func Caches() *schema.Table {
 		Multiplex:   client.SubscriptionMultiplexRegisteredNamespace("azure_redis_caches", client.Namespacemicrosoft_cache),
 		Transform:   transformers.TransformWithStruct(&armredis.ResourceInfo{}),
 		Columns: []schema.Column{
+			{
+				Name:     "subscription_id",
+				Type:     schema.TypeString,
+				Resolver: client.ResolveAzureSubscription,
+			},
 			{
 				Name:     "id",
 				Type:     schema.TypeString,