+++ plugins/source/azure/resources/services/subscription/subscriptions.go
@@ -24,5 +24,8 @@ func Subscriptions() *schema.Table {
 				},
 			},
 		},
+		Relations: []*schema.Table{
+			locations(),
+		},
 	}
 }